# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Restaurant API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MenuItem {
  id: number;
  name: string;
  description: string;
  price: string;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
}

export interface CreateReservationInput {
  name: string;
  email: string;
  phone: string;
  date: string;
  time: string;
  guests: number;
  /** @nullable */
  specialRequests?: string | null;
}

export interface Reservation {
  id: number;
  userId: number;
  date: string;
  time: string;
  guests: number;
  /** @nullable */
  specialRequests?: string | null;
  status: string;
  createdAt: string;
}

export interface AvailabilityResponse {
  available: boolean;
  message: string;
}

export interface ErrorResponse {
  error: string;
}

export type CheckAvailabilityParams = {
  date: string;
  time: string;
  guests: number;
};