# lib/api-zod/src/generated/types/availabilityResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Restaurant API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AvailabilityResponse {
  available: boolean;
  message: string;
}