# lib/api-zod/src/generated/types/createReservationInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Restaurant API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateReservationInput {
  name: string;
  email: string;
  phone: string;
  date: string;
  time: string;
  guests: number;
  /** @nullable */
  specialRequests?: string | null;
}