# lib/api-zod/src/generated/types/menuItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Restaurant API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MenuItem {
  id: number;
  name: string;
  description: string;
  price: string;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
}